api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
msal
requests
azure-iot-hub
azure-eventhub
pydantic
paho-mqtt